refactor(storage): trim the structured port to its used surface
Corrects overdesign found reviewing the phase 4 write seam, before the
contract acquires more callers. Behavior, wire shapes, and the web client
are unchanged.

Remove three things with no production caller:

- `SpaceRepository.compareAndSwap`. A record-only `OrderedSpaceWriter.apply`
  is the identical operation — same version precondition, same identity-field
  refusal, same result type, which the `SpaceMutationResult = SpaceWriteResult`
  alias had been admitting. The per-Space record port is now read-only
  `SpaceRecordRepository`, matching its `record` field. Its contract suite is
  retired; the single-winner race and next-record validation cases move to the
  ordered-writer suite, where the application actually relies on them.
- `NodeRepository.readMany`. Its only implementation read and parsed every
  node in the Space regardless of how many ids were asked for, so the first
  real caller would have inherited a full-Space scan per batch.
- `SpaceDeleteResult`. No repository returned it; it was composition's own
  result, hand-written as the union of the two port results it is built from.
  It becomes `SpaceDeleteOutcome` in `storage.ts`, derived rather than
  restated.

Merge `catalog()` and `lifecycle()` into one namespace-scoped `spaces()`.
They were the same kind of object — stateless, one fresh Workspace-bound
instance per call — split along a read/write line that tracked the migration
phases rather than the domain. The split forced the World rule to be resolved
in two places and left `create` reading membership it could not name through
the port. Per-Space `space(id)` is untouched: that boundary is real, since
`create` cannot be scoped to a Space that does not exist.

Narrow `StructuredBackendKind` to `'disk'` so an adapter's self-reported kind
names only kinds that exist. The wider requestable vocabulary moves to
`profile.ts` as `RequestedStructuredKind`, which is what preserves the
actionable "not implemented yet" error for a configured sqlite or postgres.

`authoritativeInsert` and the `write-suppressed` put outcome stay, now
documented as adapter-shaped the way `duplicate-node` already was. Both exist
for Disk's in-memory deletion fence and have no portable meaning a SQL
adapter would produce; naming the shared abstraction needs a second adapter.

Split the phase-labelled Disk test files into per-adapter suites beside the
adapters they cover, so no working-process phase number persists in a
filename.
